+++ CHANGELOG.md
@@ -6,6 +6,13 @@ This document is to serve as a "what has been done" in terms of the [Roadmap](ht
 - block plugin: Aloha Block Plugin has now been greatly cleaned up and improved. Besides greatly cleaned up API and documentation, the new features include Drag/Drop, Deletion and Copy/Paste support. Now fully cross-browser (IE7, IE8, IE9, Chrome, Firefox).
 - commands.delete: fixed a bug with the delete command when contents are preceded by ignorable whitespace. also added a delete test for that.
 
+## 0.20.4 - SNAPSHOT
+
+
+## 0.20.3 - 2012/01/24
+- image-plugin: The imageplugin will now only display the crop buttons when the cropping area selection was finished. This avoids a bug in Internet Explorer 7 where the crop area could not be resized once the user entered those crop buttons.
+- core: fixed floating menu pinning with topalign behaviour (topalignOffset, horizontalOffset)
+
 ## 0.20.2 - 2012/01/19
 - image-plugin: Replaced unicode characters in the crop buttons with images to fix display issues within Internet Explorer 7.
 - core: Fixed problem where Internet Explorer 7 and jquery.store will not work with frames since it will fallback to window.name storage. We'll now use a void storage for IE7. This means that IE7 will not be able to store floating menu postion and other settings.
@@ -15,6 +22,7 @@ This document is to serve as a "what has been done" in terms of the [Roadmap](ht
 - core: fixed incorrect dependency on jquery.json-2.2 where util/json2 is needed instead and made it globally available
 - core: Removed unneeded JSON empty function definition that surpressed errors in IE
 - guides: Updated guides. They now include a directory structure explanation and a detailed release guide.
+- word contenthandler: html cleanup for empty tags, removal of spans and the paragraph numbering from TOC feature
 
 ## 0.20.1 2012/01/13